requirements.txt
python-telegram-bot[job-queue]==21.10
google-generativeai==0.8.6
google-api-python-client==2.192.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
python-dotenv==1.2.2
apscheduler
pillow==12.1.1
httpx
